Auto Aim Implemented
Playing on Easy mode will aim the bat automatically for the player, letting them focus on the timing.

There should be a tweak to Batter-End Animation File that adds a property for the circle collider on the hitbox that should be off for the whole animation.
(Previous bug where the hitbox stays active during the batter-end animation allowing players to still hit the ball if they swing too early)